requirements.txt
streamlit>=1.36
pandas>=2.2
altair>=5
gspread>=6
google-auth>=2.33
extra-streamlit-components>=0.1.64
bcrypt>=4
requests>=2.32
streamlit-recaptcha>=0.0.4
cryptography
PyYAML
google-api-python-client
google-auth
google-auth-httplib2
